fix-management-useractionlogs

+++ src/main/java/cn/authing/sdk/java/dto/GetUserActionLogsDto.java
@@ -1,9 +1,6 @@
 package cn.authing.sdk.java.dto;
 
-import java.util.List;
 import com.fasterxml.jackson.annotation.JsonProperty;
-
-import cn.authing.sdk.java.dto.ListWebhooksDto;
 
 public class GetUserActionLogsDto {
     /**
@@ -35,12 +32,12 @@ public class GetUserActionLogsDto {
      * 开始时间戳
      */
     @JsonProperty("start")
-    private Integer start;
+    private Long start;
     /**
      * 结束时间戳
      */
     @JsonProperty("end")
-    private Integer end;
+    private Long end;
     /**
      * 请求是否成功
      */
@@ -87,17 +84,17 @@ public class GetUserActionLogsDto {
         this.appId = appId;
     }
 
-    public Integer getStart() {
+    public Long getStart() {
         return start;
     }
-    public void setStart(Integer start) {
+    public void setStart(Long start) {
         this.start = start;
     }
 
-    public Integer getEnd() {
+    public Long getEnd() {
         return end;
     }
-    public void setEnd(Integer end) {
+    public void setEnd(Long end) {
         this.end = end;
     }
